refactor: 🔨 Move every commands in a commands folder to reduce the pollution in the src directory
