+++ src/cli.rs
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::{generator::Generator, traces::Traces};
+use crate::commands::{generator::Generator, traces::Traces};
 use clap::{Parser, Subcommand};
 
 #[derive(Parser)]
