+++ src/commands.rs
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: 2023 The MalwareTracesGenerator development team
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+pub mod generator;
+pub mod traces;
