+++ src/commands/generator.rs
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::traces::{Runnable, Traces};
+use crate::commands::traces::{Runnable, Traces};
 use clap::Parser;
 use serde::Deserialize;
 use std::{error::Error, fs::read_to_string, path::PathBuf};
