+++ src/commands/traces.rs
@@ -2,30 +2,37 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::traces::{processes::spoofing::Spoofing, Runnable};
+use crate::commands::traces::{drivers::Drivers, processes::Processes};
 use clap::{Args, Subcommand};
 use serde::Deserialize;
 use std::error::Error;
 
-pub mod spoofing;
+pub mod drivers;
+pub mod processes;
 
 #[derive(Args, Deserialize)]
-pub struct Processes {
+pub struct Traces {
     #[clap(subcommand)]
     #[serde(flatten)]
     pub command: Commands,
 }
 
 #[derive(Subcommand, Deserialize)]
-#[serde(rename_all = "snake_case")]
+#[serde(rename_all = "snake_case", untagged)]
 pub enum Commands {
-    Spoofing(Spoofing),
+    Drivers(Drivers),
+    Processes(Processes),
 }
 
-impl Runnable for Processes {
+pub trait Runnable {
+    fn run(&self) -> Result<(), Box<dyn Error>>;
+}
+
+impl Runnable for Traces {
     fn run(&self) -> Result<(), Box<dyn Error>> {
         match &self.command {
-            Commands::Spoofing(spoofing) => spoofing as &dyn Runnable,
+            Commands::Drivers(drivers) => drivers as &dyn Runnable,
+            Commands::Processes(processes) => processes,
         }
         .run()
     }
