+++ src/commands/traces/drivers.rs
@@ -2,37 +2,30 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::traces::{drivers::Drivers, processes::Processes};
+use crate::commands::traces::{drivers::byovd::Byovd, Runnable};
 use clap::{Args, Subcommand};
 use serde::Deserialize;
 use std::error::Error;
 
-pub mod drivers;
-pub mod processes;
+pub mod byovd;
 
 #[derive(Args, Deserialize)]
-pub struct Traces {
+pub struct Drivers {
     #[clap(subcommand)]
     #[serde(flatten)]
     pub command: Commands,
 }
 
 #[derive(Subcommand, Deserialize)]
-#[serde(rename_all = "snake_case", untagged)]
+#[serde(rename_all = "snake_case")]
 pub enum Commands {
-    Drivers(Drivers),
-    Processes(Processes),
+    Byovd(Byovd),
 }
 
-pub trait Runnable {
-    fn run(&self) -> Result<(), Box<dyn Error>>;
-}
-
-impl Runnable for Traces {
+impl Runnable for Drivers {
     fn run(&self) -> Result<(), Box<dyn Error>> {
         match &self.command {
-            Commands::Drivers(drivers) => drivers as &dyn Runnable,
-            Commands::Processes(processes) => processes,
+            Commands::Byovd(byovd) => byovd as &dyn Runnable,
         }
         .run()
     }
