+++ src/commands/traces/drivers/byovd.rs
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::{traces::Runnable, windows::users::is_administrator};
+use crate::{commands::traces::Runnable, windows::users::is_administrator};
 use clap::Parser;
 use serde::Deserialize;
 use std::{error::Error, path::PathBuf};
