+++ src/commands/traces/processes.rs
@@ -2,15 +2,15 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::traces::{drivers::byovd::Byovd, Runnable};
+use crate::commands::traces::{processes::spoofing::Spoofing, Runnable};
 use clap::{Args, Subcommand};
 use serde::Deserialize;
 use std::error::Error;
 
-pub mod byovd;
+pub mod spoofing;
 
 #[derive(Args, Deserialize)]
-pub struct Drivers {
+pub struct Processes {
     #[clap(subcommand)]
     #[serde(flatten)]
     pub command: Commands,
@@ -19,13 +19,13 @@ pub struct Drivers {
 #[derive(Subcommand, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum Commands {
-    Byovd(Byovd),
+    Spoofing(Spoofing),
 }
 
-impl Runnable for Drivers {
+impl Runnable for Processes {
     fn run(&self) -> Result<(), Box<dyn Error>> {
         match &self.command {
-            Commands::Byovd(byovd) => byovd as &dyn Runnable,
+            Commands::Spoofing(spoofing) => spoofing as &dyn Runnable,
         }
         .run()
     }
