+++ src/commands/traces/processes/spoofing.rs
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::{traces::Runnable, windows::processes::get_pid};
+use crate::{commands::traces::Runnable, windows::processes::get_pid};
 use clap::Parser;
 use serde::Deserialize;
 use std::{
