+++ src/main.rs
@@ -3,14 +3,13 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 mod cli;
-mod generator;
-mod traces;
+mod commands;
 mod windows;
 
 use clap::Parser;
 use cli::{Arguments, Commands};
+use commands::traces::Runnable;
 use std::error::Error;
-use traces::Runnable;
 
 fn main() -> Result<(), Box<dyn Error>> {
     match Arguments::parse().command {
